+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Cluster specific agents API (`/cluster/:id/agents`).
 - Cluster specific events API (`/cluster/:id/events`).
 - HTTP Discovery.